requirements.txt
easyocr>=1.7.0
Pillow>=9.0.0
numpy>=1.21.0
gspread>=5.0.0
oauth2client>=4.1.3
torch>=1.12.0
torchvision>=0.13.0